Fix icons page
